+++ src/main/java/br/unitins/tp1/roteadores/service/pedido/LoteServiceImpl.java
@@ -6,6 +6,7 @@ import br.unitins.tp1.roteadores.dto.pedido.LoteRequestDTO;
 import br.unitins.tp1.roteadores.model.pedido.Lote;
 import br.unitins.tp1.roteadores.repository.LoteRepository;
 import br.unitins.tp1.roteadores.service.roteador.RoteadorService;
+import br.unitins.tp1.roteadores.validation.ValidationException;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
@@ -21,11 +22,17 @@ public class LoteServiceImpl implements LoteService {
 
     @Override
     public Lote findById(Long id) {
+        if (loteRepository.findById(id) == null)
+            throw new ValidationException("id", "id nao encontrado");
+        
         return loteRepository.findById(id);
     }
 
     @Override
     public Lote findByIdRoteador(Long idRoteador) {
+        if (loteRepository.findByIdRoteador(idRoteador) == null)
+            throw new ValidationException("idRoteador", "id do roteador nao encontrado.");
+
         return loteRepository.findByIdRoteador(idRoteador);
     }
 
@@ -36,6 +43,9 @@ public class LoteServiceImpl implements LoteService {
 
     @Override
     public Lote findByCodigo(String codigo) {
+        if (loteRepository.findByCodigo(codigo) == null)
+            throw new ValidationException("codigo", "Codigo nao encontrado");
+        
         return loteRepository.findByCodigo(codigo);
     } 
 
@@ -47,6 +57,9 @@ public class LoteServiceImpl implements LoteService {
     @Override
     @Transactional
     public Lote create(LoteRequestDTO dto) {
+        if (loteRepository.findByIdRoteador(dto.idRoteador()) == null)
+            throw new ValidationException("idRoteador", "id do roteador nao encontrado.");
+
         // buscando o estado a partir de um id do lote
         Lote lote = new Lote();
         lote.setRoteador(roteadorService.findById(dto.idRoteador()));
@@ -63,6 +76,12 @@ public class LoteServiceImpl implements LoteService {
     @Override
     @Transactional
     public Lote update(Long id, LoteRequestDTO dto) {
+        if (loteRepository.findById(id) == null)
+            throw new ValidationException("id", "id nao encontrado");
+
+        if (loteRepository.findByIdRoteador(dto.idRoteador()) == null)
+            throw new ValidationException("idRoteador", "id do roteador nao encontrado.");
+
         Lote lote = loteRepository.findById(id);
 
         lote.setRoteador(roteadorService.findById(dto.idRoteador()));
@@ -76,6 +95,9 @@ public class LoteServiceImpl implements LoteService {
     @Override
     @Transactional
     public void delete(Long id) {
+        if (loteRepository.findById(id) == null)
+            throw new ValidationException("id", "id nao encontrado");
+            
         loteRepository.deleteById(id);
     }
